Refatorando /pages e adicionando redirecionamento de paginas
